requirements.txt
h5py==2.9.0
numpy==1.16.2
opencv-python==4.1.0.25
tensorboard==1.12.0
tensorflow-gpu==1.12.0
tqdm==4.31.1

